TP-1174 Fixed paragraph table key errors

+++ public/modules/custom/hel_tpm_file_garbage_collector/src/Plugin/QueueWorker/FileGarbageWorker.php
@@ -50,6 +50,9 @@ final class FileGarbageWorker extends QueueWorkerBase implements ContainerFactor
   public function processItem($data): void {
     $storage = $this->entityTypeManager->getStorage('file');
     $file = $storage->load($data['fid']);
+    if (empty($file)) {
+      return;
+    }
     $file->delete();
   }
